organize components into files
