+++ packages/player/src/__tests__/tsconfig.json
@@ -1,4 +1,7 @@
 {
   "extends": "../../tsconfig.json",
-  "include": [".", "../."]
+  "include": [
+    ".",
+    "../."
+  ]
 }
