+++ packages/player/src/styles/scss/components/bezel.scss
@@ -7,7 +7,7 @@
   z-index: 17;
   margin-left: -26px;
   margin-top: -26px;
-  background: rgba(0,0,0,.5);
+  background: rgba(0, 0, 0, .5);
   border-radius: 26px;
 
   &.cueplayer-react-bezel-animation {
