+++ packages/player/src/styles/scss/components/button.scss
@@ -13,6 +13,6 @@
   transition: none;
 
   -webkit-appearance: none;
-  -moz-appearance:    none;
-  appearance:         none;
+  -moz-appearance: none;
+  appearance: none;
 }
