+++ packages/player/src/styles/scss/components/control.scss
@@ -23,11 +23,10 @@
   &:hover:before,
   &:focus {
     text-shadow: 0em 0em 1em $cueplayer-react-primary-foreground-color,
-      0em 0em 0.5em $cueplayer-react-primary-foreground-color;
+    0em 0em 0.5em $cueplayer-react-primary-foreground-color;
   }
 
 }
-
 
 
 // Hide control text visually, but have it available for screenreaders
