+++ packages/player/src/styles/scss/components/loading-spinner.scss
@@ -44,12 +44,12 @@
 
 .cueplayer-react-seeking .cueplayer-react-loading-spinner:before,
 .cueplayer-react-waiting .cueplayer-react-loading-spinner:before {
-  border-top-color: rgb(255,255,255);
+  border-top-color: rgb(255, 255, 255);
 }
 
 .cueplayer-react-seeking .cueplayer-react-loading-spinner:after,
 .cueplayer-react-waiting .cueplayer-react-loading-spinner:after {
-  border-top-color: rgb(255,255,255);
+  border-top-color: rgb(255, 255, 255);
   -webkit-animation-delay: 0.44s;
   animation-delay: 0.44s;
 }
