+++ packages/player/src/styles/scss/mixins.scss
@@ -1,4 +1,3 @@
-
 @mixin hide-visually {
   border: 0;
   clip: rect(0 0 0 0);
@@ -83,7 +82,7 @@
 }
 
 // https://developer.mozilla.org/en-US/docs/Web/CSS/box-shadow
-@mixin box-shadow ($string: 0 0 1em rgba(0, 0, 0, 0.25)) {
+@mixin box-shadow($string: 0 0 1em rgba(0, 0, 0, 0.25)) {
   -webkit-box-shadow: $string;
   -moz-box-shadow: $string;
   box-shadow: $string;
